requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
django-extensions==3.2.1
django-filter==22.1
django-stubs-ext==0.7.0; python_version >= '3.7'
django-stubs==1.13.0
django==4.1.3
djangorestframework==3.14.0
markdown==3.4.1
mypy-extensions==0.4.3
mypy==0.982; python_version >= '3.7'
mysqlclient==2.1.1
pytz==2022.6
sqlparse==0.4.3; python_version >= '3.5'
tomli==2.0.1; python_version >= '3.7'
types-pytz==2022.6.0.1
types-pyyaml==6.0.12.1
typing-extensions==4.4.0; python_version >= '3.7'
